Move Curl_get_line from cookie.c to a generic file
It's a better way to fix build with ALTSVC enabled and COOKIES disabled.
